# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarberMetrics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: string;
  name: string;
  price: number;
  durationMinutes?: number | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateServiceBody {
  name: string;
  price: number;
  durationMinutes?: number | null;
}

export interface UpdateServiceBody {
  name?: string;
  price?: number;
  durationMinutes?: number | null;
  isActive?: boolean;
}

export interface Appointment {
  id: string;
  serviceId?: string | null;
  serviceName: string;
  price: number;
  startedAt: string;
  endedAt?: string | null;
  durationSeconds?: number | null;
  note?: string | null;
  createdAt: string;
}

export interface CreateAppointmentBody {
  serviceId?: string | null;
  serviceName: string;
  price: number;
  startedAt: string;
  endedAt?: string | null;
  durationSeconds?: number | null;
  note?: string | null;
}

export interface UpdateAppointmentBody {
  serviceId?: string | null;
  serviceName?: string;
  price?: number;
  startedAt?: string;
  endedAt?: string | null;
  durationSeconds?: number | null;
  note?: string | null;
}

export interface DailySummary {
  date: string;
  totalRevenue: number;
  totalAppointments: number;
  avgTicket: number;
  goalProgressPct: number;
  goal: number;
  workedSeconds: number;
  idleSeconds: number;
  revenuePerHour: number;
}

export interface DailyPoint {
  date: string;
  revenue: number;
  appointments: number;
  workedSeconds: number;
}

export interface RangeSummary {
  startDate: string;
  endDate: string;
  totalRevenue: number;
  totalAppointments: number;
  avgTicket: number;
  workedSeconds: number;
  idleSeconds: number;
  revenuePerHour: number;
  days: DailyPoint[];
}

export interface Insight {
  id: string;
  /** e.g. projection, comparison, trend */
  kind: string;
  title: string;
  body: string;
  /** positive, neutral, warning */
  tone: string;
}

export interface Settings {
  barbershopName: string;
  dailyGoal: number;
  currency: string;
  workStartTime: string;
  workEndTime: string;
  workDays: number[];
  theme: string;
}

export interface UpdateSettingsBody {
  barbershopName?: string;
  dailyGoal?: number;
  currency?: string;
  workStartTime?: string;
  workEndTime?: string;
  workDays?: number[];
  theme?: string;
}

export type ListAppointmentsParams = {
  /**
   * A specific YYYY-MM-DD day (in barber's local time)
   */
  date?: string;
  startDate?: string;
  endDate?: string;
  serviceId?: string;
};

export type GetDailySummaryParams = {
  date: string;
};

export type GetRangeSummaryParams = {
  startDate: string;
  endDate: string;
};

export type GetInsightsParams = {
  date: string;
};