# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BarberMetrics API
servers:
  - url: /api
    description: Base API path
tags:
  - name: services
    description: Catalog of services offered by the barber
  - name: appointments
    description: Appointments / cash book
  - name: summary
    description: Aggregated stats and insights
  - name: settings
    description: Profile and settings
  - name: health
    description: Health operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List active services in the catalog
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a new service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments, optionally filtered by date or range
      parameters:
        - in: query
          name: date
          required: false
          description: A specific YYYY-MM-DD day (in barber's local time)
          schema:
            type: string
        - in: query
          name: startDate
          required: false
          schema:
            type: string
        - in: query
          name: endDate
          required: false
          schema:
            type: string
        - in: query
          name: serviceId
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create a new appointment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /summary/daily:
    get:
      operationId: getDailySummary
      tags: [summary]
      summary: Daily summary (revenue, appointments, avg ticket, goal progress)
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySummary"
  /summary/range:
    get:
      operationId: getRangeSummary
      tags: [summary]
      summary: Daily aggregates between start and end date (inclusive)
      parameters:
        - in: query
          name: startDate
          required: true
          schema:
            type: string
        - in: query
          name: endDate
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RangeSummary"
  /summary/insights:
    get:
      operationId: getInsights
      tags: [summary]
      summary: Predictive insights derived from the last 7-30 days
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insight"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get profile and settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update profile and settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Service:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, price, isActive, createdAt]
    CreateServiceBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
          nullable: true
      required: [name, price]
    UpdateServiceBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
          nullable: true
        isActive:
          type: boolean

    Appointment:
      type: object
      properties:
        id:
          type: string
        serviceId:
          type: string
          nullable: true
        serviceName:
          type: string
        price:
          type: number
        startedAt:
          type: string
        endedAt:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        note:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, serviceName, price, startedAt, createdAt]
    CreateAppointmentBody:
      type: object
      properties:
        serviceId:
          type: string
          nullable: true
        serviceName:
          type: string
        price:
          type: number
        startedAt:
          type: string
        endedAt:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        note:
          type: string
          nullable: true
      required: [serviceName, price, startedAt]
    UpdateAppointmentBody:
      type: object
      properties:
        serviceId:
          type: string
          nullable: true
        serviceName:
          type: string
        price:
          type: number
        startedAt:
          type: string
        endedAt:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        note:
          type: string
          nullable: true

    DailySummary:
      type: object
      properties:
        date:
          type: string
        totalRevenue:
          type: number
        totalAppointments:
          type: integer
        avgTicket:
          type: number
        goalProgressPct:
          type: number
        goal:
          type: number
        workedSeconds:
          type: integer
        idleSeconds:
          type: integer
        revenuePerHour:
          type: number
      required:
        - date
        - totalRevenue
        - totalAppointments
        - avgTicket
        - goalProgressPct
        - goal
        - workedSeconds
        - idleSeconds
        - revenuePerHour

    DailyPoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        appointments:
          type: integer
        workedSeconds:
          type: integer
      required: [date, revenue, appointments, workedSeconds]
    RangeSummary:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        totalRevenue:
          type: number
        totalAppointments:
          type: integer
        avgTicket:
          type: number
        workedSeconds:
          type: integer
        idleSeconds:
          type: integer
        revenuePerHour:
          type: number
        days:
          type: array
          items:
            $ref: "#/components/schemas/DailyPoint"
      required:
        - startDate
        - endDate
        - totalRevenue
        - totalAppointments
        - avgTicket
        - workedSeconds
        - idleSeconds
        - revenuePerHour
        - days

    Insight:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
          description: e.g. projection, comparison, trend
        title:
          type: string
        body:
          type: string
        tone:
          type: string
          description: positive, neutral, warning
      required: [id, kind, title, body, tone]

    Settings:
      type: object
      properties:
        barbershopName:
          type: string
        dailyGoal:
          type: number
        currency:
          type: string
        workStartTime:
          type: string
        workEndTime:
          type: string
        workDays:
          type: array
          items:
            type: integer
        theme:
          type: string
      required:
        - barbershopName
        - dailyGoal
        - currency
        - workStartTime
        - workEndTime
        - workDays
        - theme
    UpdateSettingsBody:
      type: object
      properties:
        barbershopName:
          type: string
        dailyGoal:
          type: number
        currency:
          type: string
        workStartTime:
          type: string
        workEndTime:
          type: string
        workDays:
          type: array
          items:
            type: integer
        theme:
          type: string