# lib/api-zod/src/generated/types/dailyPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarberMetrics API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyPoint {
  date: string;
  revenue: number;
  appointments: number;
  workedSeconds: number;
}