# lib/api-zod/src/generated/types/dailySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarberMetrics API
 * OpenAPI spec version: 0.1.0
 */

export interface DailySummary {
  date: string;
  totalRevenue: number;
  totalAppointments: number;
  avgTicket: number;
  goalProgressPct: number;
  goal: number;
  workedSeconds: number;
  idleSeconds: number;
  revenuePerHour: number;
}