# lib/api-zod/src/generated/types/rangeSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarberMetrics API
 * OpenAPI spec version: 0.1.0
 */
import type { DailyPoint } from "./dailyPoint";

export interface RangeSummary {
  startDate: string;
  endDate: string;
  totalRevenue: number;
  totalAppointments: number;
  avgTicket: number;
  workedSeconds: number;
  idleSeconds: number;
  revenuePerHour: number;
  days: DailyPoint[];
}