# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarberMetrics API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: string;
  name: string;
  price: number;
  durationMinutes?: number | null;
  isActive: boolean;
  createdAt: string;
}