# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarberMetrics API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  barbershopName: string;
  dailyGoal: number;
  currency: string;
  workStartTime: string;
  workEndTime: string;
  workDays: number[];
  theme: string;
}