# lib/api-zod/src/generated/types/updateAppointmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarberMetrics API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAppointmentBody {
  serviceId?: string | null;
  serviceName?: string;
  price?: number;
  startedAt?: string;
  endedAt?: string | null;
  durationSeconds?: number | null;
  note?: string | null;
}