# lib/api-zod/src/generated/types/updateServiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarberMetrics API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateServiceBody {
  name?: string;
  price?: number;
  durationMinutes?: number | null;
  isActive?: boolean;
}